Refactored Scheduler component, Created ScheduleEntryContextMenu Component, Created ScheduleEntryDialog component
